Align `swift sdk` configuration subcommand with proposal text
Per [the proposal text](https://github.com/apple/swift-evolution/blob/main/proposals/0387-cross-compilation-destinations.md#swift-sdk-installation-and-configuration):

> swift sdk configure <identifier> <target-triple>`, which allows users to provide additional search paths and toolsets to be used subsequently when building with a given Swift SDK. Specifically, multiple `--swift-resources-path`, `--include-search-path`, `--library-search-path`, and `--toolset` options with corresponding paths can be provided, which then will be stored as configuration for this Swift SDK. `swift sdk configure <identifier> --show-configuration` will print currently set paths, while `swift sdk configure <identifier> --reset` will reset all of those at once.
